requirements.txt
streamlit==1.32.0
pandas
plotly
openpyxl
google-cloud-bigquery
db-dtypes